requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.3.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
urllib3==1.25.8
websocket-client==0.48.0
Werkzeug==1.0.1
